Remove commented layer descriptions; add new function layer skeleton
Comments are often lies. That holds true here, as I was not updating
them as I went along.

I tried to make this two commits, but my patch command wouldn't split
them correctly and I figured that it was more trouble than it was worth
for keyboard config commits :sweat_smile_emoji: