Add a sticky hyper layer, activated from LOWER(H)
This layer uses sticky key chains to send the default layer code from
that key position but with all `hyper` modifiers also pressed down.

From what I can tell in the documentation, this behavior may be the same
as using the `&kp` behavior in place of `&sk` to send the chained keys
together. Will test this in a subsequent commit, since technically I am
not using sticky keys here the way that they are intended.